fix: no migrations on worker and provider startup

+++ packages/projection-typeorm/src/createDataSource.ts
@@ -16,7 +16,7 @@ type PostgresConnectionOptions = DataSourceOptions & { type: 'postgres' };
 
 export type PgConnectionConfig = Pick<
   PostgresConnectionOptions,
-  'host' | 'port' | 'database' | 'username' | 'password' | 'ssl'
+  'host' | 'port' | 'database' | 'username' | 'password' | 'ssl' | 'poolSize'
 >;
 
 export type TypeormDevOptions = Pick<PostgresConnectionOptions, 'synchronize' | 'dropSchema'>;
@@ -149,12 +149,14 @@ export const createDataSource = ({
   return patchObject(dataSource, {
     async initialize() {
       await dataSource.initialize();
-      await initializePgBoss(
-        dataSource,
-        contextLogger(logger, 'createDataSource'),
-        extensions?.pgBoss,
-        devOptions?.dropSchema
-      );
+      if (extensions?.pgBoss && (options?.migrationsRun || devOptions?.synchronize)) {
+        await initializePgBoss(
+          dataSource,
+          contextLogger(logger, 'createDataSource'),
+          extensions?.pgBoss,
+          devOptions?.dropSchema
+        );
+      }
       return dataSource;
     }
   });
